+++ contrib/brl/bseg/boxm2/reg/ocl/boxm2_ocl_hierarchical_reg.cxx
@@ -50,11 +50,11 @@ bool boxm2_ocl_hierarchical_reg::exhaustive()
             int searchwidth = numsamples_search_width[level];
             vnl_vector<double> best_sample = iter->second;
             int numsamples_per_best_particle = (int) std::pow((float)searchwidth,(float)params_to_vary);
-            for(unsigned int sampleno = 0 ; sampleno < numsamples_per_best_particle; sampleno++)
+            for(size_t sampleno = 0 ; sampleno < numsamples_per_best_particle; sampleno++)
             {
                 int cont=sampleno;
                 vnl_vector<double> curr_sample = best_sample;
-                for(unsigned k = 0; k <params_to_vary; k++)
+                for(size_t k = 0; k <params_to_vary; k++)
                 {
                     int t = params_to_vary - k - 1;
                     unsigned int var = cont/(unsigned int)std::pow((float)searchwidth,(float)t); // quotient
